refactor(holdings): extract duplicated price-loading helpers into shared BacktestPriceLoader
